#821 [REFACTOR] store를 shared로 이동
