+++ src/shared/store/index.js
